requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.7.6; python_version >= '3.6'
bidict==0.21.4; python_version >= '3.6'
cffi==1.15.0; platform_python_implementation == 'CPython' and sys_platform == 'win32'
click==8.0.4; python_version >= '3.6'
colorama==0.4.4; platform_system == 'Windows'
flask-cors==3.0.10
flask-jwt-extended==4.3.1
flask-marshmallow==0.14.0
flask-migrate==3.1.0
flask-socketio==5.1.1
flask-sqlalchemy==2.5.1
flask==2.0.3
gevent==21.12.0
greenlet==1.1.2; platform_python_implementation == 'CPython'
gunicorn==20.1.0
h11==0.13.0; python_version >= '3.6'
itsdangerous==2.1.0; python_full_version >= '3.7.0'
jinja2==3.0.3; python_version >= '3.6'
mako==1.1.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.1.0; python_full_version >= '3.7.0'
marshmallow-sqlalchemy==0.27.0
marshmallow==3.14.1; python_version >= '3.6'
psycopg2==2.9.3
pycparser==2.21
pyjwt==2.3.0; python_version >= '3.6'
python-decouple==3.6
python-engineio==4.3.1; python_version >= '3.6'
python-socketio==5.5.2; python_version >= '3.6'
setuptools==60.9.3; python_full_version >= '3.7.0'
simple-websocket==0.5.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.4.31; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
werkzeug==2.0.3; python_version >= '3.6'
wsproto==1.1.0; python_full_version >= '3.7.0'
zope.event==4.5.0
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
